@@ -4,13 +4,13 @@ version: 0.1.8
 homepage: https://github.com/eslamfaisal/FlutterCellInfo
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
-  flutter: ">=1.20.0"
+  sdk: '>=2.12.0 <4.0.0'
+  flutter: ">=2.12.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  flare_flutter: ^2.0.3
+  flare_flutter: ^3.0.2
 
 dev_dependencies:
   flutter_test:
